# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OrgFlow API
  version: 1.0.0
  description: Employee and leave management. All routes except login and password reset need a Bearer JWT.
servers:
  - url: http://localhost:8080/api
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    LeaveRequest:
      type: object
      properties:
        id: { type: integer }
        employeeName: { type: string }
        leaveType: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        days: { type: integer, description: Working days (weekends and company holidays excluded) }
        status: { type: string, enum: [pending, approved, rejected, cancelled] }
        attachments: { type: array, items: { type: object, properties: { id: { type: integer }, name: { type: string }, size: { type: integer } } } }
security:
  - bearer: []
paths:
...
  /leave/calendar:
    get:
      summary: Approved leave and company holidays for a month
      parameters: [{ name: month, in: query, schema: { type: string, example: 2026-11 } }]
      responses: { '200': { description: "{ month, leave[], holidays[] }" } }
  /leave/export.csv:
    get:
      summary: Leave report as CSV (admin)
      parameters: [{ name: year, in: query, schema: { type: integer } }]
      responses: { '200': { description: text/csv } }
  /leave:
    get:
      summary: List requests
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, team, all] } }
        - { name: status, in: query, schema: { type: string } }
      responses: { '200': { description: list, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/LeaveRequest' } } } } } }
    post:
      summary: Request leave (checks overlap, balance, working days)
      responses: { '201': { description: created }, '409': { description: overlaps an existing request }, '422': { description: insufficient balance } }
  /leave/{id}/review:
    patch: { summary: Approve or reject (the employee's manager or an admin; never your own request), parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: reviewed }, '409': { description: already decided } } }
  /leave/{id}/cancel:
    patch: { summary: Cancel your own pending request, parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: cancelled } } }
  /leave/{id}/attachments:
    post: { summary: "Upload PDF/PNG/JPG (max 5 MB, 3 per request), multipart field 'file'", parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '201': { description: uploaded }, '413': { description: too large } } }
  /leave/{id}/attachments/{fileId}:
    get: { summary: Download (requester, their manager, admin), parameters: [{ name: id, in: path, required: true, schema: { type: integer } }, { name: fileId, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: file } } }
    delete: { summary: Remove while pending (requester), parameters: [{ name: id, in: path, required: true, schema: { type: integer } }, { name: fileId, in: path, required: true, schema: { type: integer } }], responses: { '204': { description: removed } } }
...